+++ feature/viewit/src/main/java/com/teamwable/viewit/ui/ViewItScreen.kt
@@ -14,6 +14,7 @@ import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
@@ -33,16 +34,22 @@ import com.teamwable.designsystem.component.paging.WableCustomRefreshIndicator
 import com.teamwable.designsystem.component.paging.WablePagingSpinner
 import com.teamwable.designsystem.component.screen.LoadingScreen
 import com.teamwable.designsystem.component.screen.NewsNoticeEmptyScreen
+import com.teamwable.designsystem.extension.composable.scrollToTop
 import com.teamwable.designsystem.theme.WableTheme
 import com.teamwable.designsystem.type.ContentType
 import com.teamwable.model.viewit.ViewIt
-import com.teamwable.ui.extensions.awaitRefreshComplete
 import com.teamwable.ui.type.BottomSheetType
 import com.teamwable.ui.type.SnackbarType
 import com.teamwable.viewit.R
 import com.teamwable.viewit.component.ViewitItem
 import com.teamwable.viewit.viewit.ViewItViewModel
 import kotlinx.coroutines.flow.collectLatest
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.filter
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.timeout
+import kotlin.time.Duration.Companion.seconds
 
 @Composable
 fun ViewItRoute(
@@ -77,7 +84,7 @@ fun ViewItRoute(
                     ViewItSideEffect.UI.Refresh -> {
                         viewIts.refresh()
                         awaitRefreshComplete(viewIts)
-                        listState.animateScrollToItem(0, -100)
+                        listState.scrollToTop()
                     }
                 }
             }
@@ -100,6 +107,15 @@ fun rememberViewItActions(
             onRefresh = { viewModel.onIntent(ViewItIntent.PullToRefresh) },
         )
     }
+}
+
+suspend fun awaitRefreshComplete(data: LazyPagingItems<*>) {
+    snapshotFlow { data.loadState.refresh }
+        .map { it is LoadState.NotLoading }
+        .distinctUntilChanged()
+        .filter { it }
+        .timeout(5.seconds)
+        .first()
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
